Changed int to float for MO

+++ examples/cannyStencil/cannyThreshold.cl
@@ -1,6 +1,6 @@
 float func(input_matrix_t* img, __global float* kernelVec, int range)
 {
-    int val = getData(img, 0, 0);
+    float val = getData(img, 0, 0);
     if(val <= 20) {
         val = 0;
     } else if(val > 20 && val < 40) {
